refactor: :truck: description list app 라우터로 변경
