+++ src/components/DescriptionList/descriptionList.module.scss
@@ -1,0 +1,85 @@
+@use '@/styles/constants/colors';
+@use '@/styles/constants/sizes';
+@use '@/styles/responsive';
+
+.descriptionListWrapper {
+  padding: 10px;
+  width: 100%;
+  max-width: sizes.$MAIN_INNER_MAX_WIDTH;
+}
+
+.descriptionElement {
+  margin-bottom: 20px;
+
+  h5 {
+    font-size: sizes.$FONT_ML;
+    color: colors.$GRAY3;
+    margin-bottom: 10px;
+  }
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+}
+
+.descriptionInnerList {
+  padding-left: 25;
+
+  li {
+    list-style: square;
+    margin-bottom: 5px;
+  }
+
+  li:last-child {
+    margin-bottom: 0;
+  }
+}
+
+.descriptionParagraph {
+  word-break: keep-all;
+  width: 100%;
+  white-space: pre-line;
+  font-size: sizes.$FONT_M;
+  color: colors.$GRAY6;
+  line-height: calc(sizes.$FONT_M + 8px);
+}
+
+.descriptionLinkWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding-left: 25px;
+  margin-top: 10px;
+
+  svg {
+    width: 15px;
+    height: 15px;
+    fill: colors.$GRAYA;
+    margin-right: 5px;
+    transition: color 0.2s ease-in;
+  }
+
+  li {
+    padding-bottom: 10px;
+  }
+
+  li:last-child {
+    padding-bottom: 0px;
+  }
+
+  span {
+    color: colors.$GRAYA;
+    font-size: sizes.$FONT_MS;
+    transition: color 0.2s ease-in;
+  }
+
+  &:hover {
+    span {
+      color: colors.$GRAY9;
+    }
+
+    svg {
+      fill: colors.$GRAY9;
+    }
+  }
+}
