+++ src/components/ProjectComponent/index.tsx
@@ -2,15 +2,13 @@ import { ProjectType } from '@/types/profileData';
 import { ProjectInfoHeader } from './ProjectInfoHeader';
 
 interface Props {
-  children: JSX.Element;
   project: ProjectType;
 }
 
-export const ProjectComponent = ({ children, project }: Props) => {
+export const ProjectComponent = ({ project }: Props) => {
   return (
-    <div>
+    <>
       <ProjectInfoHeader project={project} />
-      {children}
-    </div>
+    </>
   );
 };
